refactor aws, swagger and security configs, carsn and security services. Fix tests

+++ controller/src/main/java/com/implemica/controller/service/amazonS3/AmazonClient.java
@@ -1,7 +1,6 @@
 package com.implemica.controller.service.amazonS3;
 
 import com.amazonaws.services.s3.AmazonS3;
-
 import com.amazonaws.services.s3.model.CannedAccessControlList;
 import com.amazonaws.services.s3.model.DeleteObjectRequest;
 import com.amazonaws.services.s3.model.ObjectMetadata;
@@ -9,6 +8,7 @@ import com.amazonaws.services.s3.model.PutObjectRequest;
 import com.implemica.controller.exceptions.InvalidImageTypeException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
@@ -16,6 +16,7 @@ import java.io.File;
 import java.util.Objects;
 
 import static com.implemica.controller.utils.ConverterDTO.convertMultiPartToFile;
+import static java.lang.String.format;
 
 @Service
 @Slf4j
@@ -24,19 +25,21 @@ public class AmazonClient {
 
     private final AmazonS3 s3client;
 
-//    private final static String bucketName = "carcatalogcarsphotop";
-    private final static String bucketName = "cars-storage-yaroslav-b.implemica.com";
+    private static String bucketName;
 
+    private static final String CACHE_CONTROL_HEADER_PREFIX = "max-age";
+
+    @Value("${cars.s3.images_cahcecontrol}")
+    private int cacheControlDuration;
 
     public void uploadFileTos3bucket(String fileName, MultipartFile multipartFile) throws InvalidImageTypeException {
 
         log.info("bucket name : " + bucketName);
-        log.info("region : " + s3client.getRegion());
 
         checkImageType(multipartFile);
 
         ObjectMetadata objectMetadata = new ObjectMetadata();
-        objectMetadata.setCacheControl("max-age=1209600");
+        objectMetadata.setCacheControl(format("%s=%d", CACHE_CONTROL_HEADER_PREFIX, cacheControlDuration));
 
         File file = convertMultiPartToFile(multipartFile);
 
@@ -48,7 +51,11 @@ public class AmazonClient {
 
     public void deleteFileFromS3Bucket(String imageName) {
 
-        s3client.deleteObject(new DeleteObjectRequest( bucketName, imageName));
+        if(!imageName.equals("default-car-image")){
+
+            s3client.deleteObject(new DeleteObjectRequest(bucketName, imageName));
+        }
+
     }
 
     private static void checkImageType(MultipartFile image) throws InvalidImageTypeException {
@@ -60,6 +67,11 @@ public class AmazonClient {
 
             throw new InvalidImageTypeException("Invalid image type.");
         }
+    }
+
+    @Value("${cars.s3.bucket_name}")
+    public void setBucketName(String bucketName) {
+        AmazonClient.bucketName = bucketName;
     }
 
     public static String getBucketName() {
